+++ fastapi/app/api/routes/auth/jwt.py
@@ -1,12 +1,13 @@
 from jose import jwt
 from fastapi import APIRouter, HTTPException, Depends
+from fastapi.responses import JSONResponse
 
 from config import get_settings
 from services.auth import auth_user, create_token
 
 router = APIRouter(
   prefix="/jwt-test",
-  tags=["api_common_jwt-test"],
+  tags=["api_jwt-test"],
   # dependencies=[Depends(get_token_header)],
   responses={404: {"description": "Not found"}},
 )
@@ -14,14 +15,12 @@ router = APIRouter(
 @router.get("/fail")
 async def jwt_test_fail():
   try:
-    tokens = create_token({ "key": "value" })
+    tokens = await create_token({ "key": "value" })
     decoded = jwt.decode(tokens["access_token"], "secret111", algorithms=["HS256"])
-  except (jwt.JWTError):
-    raise HTTPException(
-      status_code=403,
-      detail="Could not validate credentials",
-    )
-  return { "token": token, "decoded": decoded }
+  except jwt.JWTError:
+    raise HTTPException( status_code=403, detail="Could not validate credentials" )
+    # return JSONResponse( status_code=418, content={"message": "Force Failure"} )
+  return { "access_token": tokens["access_token"], "decoded": decoded }
 
 @router.get("/")
 async def jwt_test():
@@ -32,3 +31,12 @@ async def jwt_test():
 @router.get("/test-middleware")
 async def jwt_test_middleware(decoded = Depends(auth_user)):
   return decoded
+
+# response.set_cookie(oauth2_scheme.token_name, encoded_jwt, httponly=True)
+@router.get("/fake-login")
+async def fake_login():
+  tokens = await create_token({ "key": "value" })
+  res = JSONResponse(tokens)
+  res.status_code = 200
+  # res.set_cookie(, encoded_jwt, httponly=True)
+  return res